Fix:Pay bug
